pyproject.toml
[project]
name = "diagnosis-agent"
version = "0.1.0"
description = "A log diagnosis agent"
readme = "README.md"
authors = [
    { name = "Menghao", email = "737330422@qq.com" }
]
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.66.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "uvicorn>=0.41.0",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"
